Create verfication service - Move Token Generation function out of email service
