+++ src/services/index.ts
@@ -1,2 +1,3 @@
 export * from './auth.ts';
 export * from './email.service.ts';
+export * from './verification.service.ts';
